# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "parrot api",
    "description": "鸚鸚食堂 api"
  },
  "host": "localhost:3666",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/parrot/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Parrot"
        ],
        "summary": "新增鸚鵡資料",
        "description": "新增鸚鵡資料",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "chineseName": {
                  "type": "string",
                  "example": "中文名稱"
                },
                "englishName": {
                  "type": "string",
                  "example": "英文名稱"
                },
                "chineseNickname": {
                  "type": "string",
                  "example": "中文暱稱"
                },
                "englishNickname": {
                  "type": "string",
                  "example": "英文暱稱"
                },
                "size": {
                  "type": "string",
                  "example": "體型"
                },
                "weightRange": {
                  "type": "string",
                  "example": "體重範圍(g)"
                },
                "lifespan": {
                  "type": "string",
                  "example": "預期壽命(年)"
                },
                "nativeHabitat": {
                  "type": "string",
                  "example": "原生長地"
                },
                "additionalInfo": {
                  "type": "string",
                  "example": "額外說明"
                },
                "advantages": {
                  "type": "string",
                  "example": "飼養優點"
                },
                "disadvantages": {
                  "type": "string",
                  "example": "飼養缺點"
                },
                "description": {
                  "type": "string",
                  "example": "描述"
                }
              },
              "required": [
                "chineseName",
                "englishName",
                "chineseNickname",
                "englishNickname",
                "size",
                "weightRange",
                "lifespan",
                "nativeHabitat",
                "additionalInfo",
                "advantages",
                "disadvantages",
                "description"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/feedback/": {
      "post": {
        "tags": [
          "Feedback"
        ],
        "summary": "新增回饋",
        "description": "新增回饋",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "contactPerson": {
                  "type": "string",
                  "example": "姓名"
                },
                "phone": {
                  "type": "string",
                  "example": "電話"
                },
                "email": {
                  "type": "string",
                  "example": "信箱"
                },
                "feedback": {
                  "type": "string",
                  "example": "內容"
                },
                "source": {
                  "type": "string",
                  "example": "從哪裡得知此網站"
                }
              },
              "required": [
                "contactPerson",
                "phone",
                "email",
                "feedback",
                "source"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}